move to smaller sublibrary of StochasticDiffEq and move away from extension
