+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DynamicalSystemsBase"
 uuid = "6e36e845-645a-534a-86f2-f5d4aa5a06b4"
 repo = "https://github.com/JuliaDynamics/DynamicalSystemsBase.jl.git"
-version = "3.16.0"
+version = "3.17.0"
 
 [deps]
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
@@ -14,13 +14,8 @@ SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StateSpaceSets = "40b095a5-5852-4c12-98c7-d43bf788e795"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+StochasticDiffEqHighOrder = "0520c28c-50fd-4d16-9c96-902fc80b3bab"
 SymbolicIndexingInterface = "2efcf032-c050-4f8e-a9bb-153293bab1f5"
-
-[weakdeps]
-StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
-
-[extensions]
-StochasticSystemsBase = "StochasticDiffEq"
 
 [compat]
 ForwardDiff = "0.10, 1"
@@ -30,6 +25,6 @@ Roots = "2, 3"
 SciMLBase = "3"
 StateSpaceSets = "2.5"
 Statistics = "1"
-StochasticDiffEq = "7"
+StochasticDiffEqHighOrder = "2"
 SymbolicIndexingInterface = "0.3.4"
 julia = "1.10"
